Stremlit/requirements.txt
streamlit>=1.28.0
supabase>=1.7.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
openpyxl>=3.1.0
plotly>=5.17.0
pytest>=7.4.0
pytest-cov>=4.1.0
pyspark>=3.5.0
